% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppEEARS_proc.R
\name{AppEEARS_proc}
\alias{AppEEARS_proc}
\title{Filters and gap-fills downloaded MODIS AppEEARS data}
\usage{
AppEEARS_proc(
  unpacked_LAI,
  fit_method,
  plot = FALSE,
  write_output = FALSE,
  save_dir = NULL
)
}
\arguments{
\item{unpacked_LAI}{Output from the AppEEARS_unpack_QC function}

\item{fit_method}{There are several options available from the phenofit package including
AG", "Beck", "Elmore", "Gu", "Klos", "Zhang".}

\item{write_output}{Binary indicating whether to write each individual driver
file to disk. Default value is FALSE.}

\item{save_dir}{Optional parameter when write_output = TRUE. The save directory 
for files to be placed in. For example, "C:/}

\item{Plot}{Binary, where plot = TRUE generates a plot and plot = FALSE does not}
}
\description{
This function filters and gap-fills downloaded MODIS LAI data from
the lpdaac AppEEARS download portal. There are several choices for processing
this data:

\itemize{
  \item The LAI data is processed using the phenofit R package 
  https://github.com/kongdd/phenofit. A rough fit is done using a weighted
  Whittaker smoother and then curve fitting is done using an assymetric
  Gaussian function. 
  \item The LAI data is processed by using a Savitsky-Golay filter to remove
  noise and then missing data is interpolated with a spline.
  \item The LAI data is processed by using a Savitsky-Golay filter to remove
  noise and then missing data is interpolated by fitting a double logistic curve
  using the greenbrown R package.
}
}
