% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppEEARS_unpack.R
\name{AppEEARS_unpack}
\alias{AppEEARS_unpack}
\title{Unpacks downloaded MODIS LAI data}
\usage{
AppEEARS_unpack(zip_file, zip_dir, request_sites)
}
\arguments{
\item{zip_file}{The name of the zip file. For example, "myzip.zip"}

\item{zip_dir}{The directory the zip file is located in}

\item{request_sites}{A string of Site_IDs}
}
\description{
This function unpacks downloaded MODIS LAI data from the lpdaac
AppEEARS download portal into a list with each item in the list representing
a single site. Additionally, data QC is performed
}
