% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeears_unpack_qc.R
\name{appeears_unpack_qc}
\alias{appeears_unpack_qc}
\title{Unpacks downloaded MODIS LAI data and QC information}
\usage{
appeears_unpack_qc(zip_file, zip_dir, request_sites)
}
\arguments{
\item{zip_file}{The name of the zip file. For example, "myzip.zip"}

\item{zip_dir}{The directory the zip file is located in}

\item{request_sites}{A string of Site_IDs}
}
\description{
This function unpacks downloaded MODIS LAI data from the lpdaac
AppEEARS download portal into a list with each item in the list representing
a single site.
}
