% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_light.R
\name{daily_light}
\alias{daily_light}
\title{Generates rough graphs and daily summaries of model outputs}
\usage{
daily_light(pred_dir, Site_ID, plot, return)
}
\arguments{
\item{pred_dir}{The read directory where model predictions are located, "C:/}

\item{Site_ID}{The Site_ID}

\item{plot}{Logical indicator. If plot = TRUE, then a timeseries of total daily
light is plotted (mol m-2 d-1) for both incoming and predicted light}

\item{return}{Logical indicator. If return = TRUE, then return a data frame
containing daily means and sums of both incoming and predicted light}
}
\value{
Returns daily mean (umol m-2 s-1) and total (kmol m-2 d-1) incoming and
predicted light
}
\description{
This function calculates daily mean (umol m-2 s-1) or total (kmol m-2 d-1)
incoming and predicted light
}
