% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDoubleLogBeck.R
\name{FitDoubleLogBeck}
\alias{FitDoubleLogBeck}
\title{A standalone version of the FitDoubleLogBeck function from the greenbrown package.}
\usage{
FitDoubleLogBeck(x, t = 1:length(x), tout = t, weighting = TRUE,
  hessian = FALSE, plot = FALSE, ninit = 30, ...)
}
\description{
This is a standalone version of the FitDoubleLogBeck function from
the greenbrown package.

Reference information
Forkel M, Wutzler T (2015) greenbrown - land surface phenology and trend analysis.
A package for the R software. Version 2.2, 2015-04-15, http://greenbrown.r-forge.r-project.org/
}
\keyword{internal}
