% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOBO_standardized.R
\name{HOBO_standardized}
\alias{HOBO_standardized}
\title{Generates a standardized set of light from HOBO loggers}
\usage{
HOBO_standardized(read_dir, filename, Lat, Lon, site_crs, estimate)
}
\arguments{
\item{read_dir}{The read directory for downloaded files. For example, "C:/}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}

\item{site_crs}{The coordinate reference system of the points, preferably designated
as an EPSG code. For example, the most common geographic system is WGS84 and its EPSG 
code is 4326.}

\item{estimate}{Binary ("yes", "no"). If yes, PPFD (umol m-2 s-1) is estimated
from the lux (lumens m-2) data}
}
\value{
Returns a standardized file of light data
}
\description{
This function takes observed light from HOBO loggers (in lux)
and generates a standardized set of validation data for comparing estimated light.
}
