% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAI_proc_dbl.R
\name{LAI_proc_dbl}
\alias{LAI_proc_dbl}
\title{Generate a continuous series of LAI using a Savitzky-Golay filter and double
logistic function}
\usage{
LAI_proc_dbl(Site)
}
\arguments{
\item{Site}{The Site ID}
}
\description{
This function generates a continuous series of LAI from downloaded
from the lpdaac AppEEARS download portal. Processing steps include:

\itemize{
  \item Savitsky-Golay filtering
  \item Fit a double logistic function to extract a continuous series of LAI
}
}
