% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai_proc_phenofit.R
\name{lai_proc_phenofit}
\alias{lai_proc_phenofit}
\title{Interpolates LAI into daily values}
\usage{
lai_proc_phenofit(Site, fit_method)
}
\arguments{
\item{Site}{The Site ID}

\item{fit_method}{The curve fitting method to pass onto phenofit ("AG", "Beck", 
"Elmore", "Gu", "Klos", "Zhang")}
}
\description{
This function generates a continuous series of LAI from downloaded
from the lpdaac AppEEARS download portal. This is done through the phenofit
R package https://github.com/kongdd/phenofit. A rough fit is done using a
weighted Whittaker smoother and then curve fitting is done using an assymetric
Gaussian function.

\itemize{
  \item Dongdong Kong, R package: A state-of-the-art Vegetation Phenology 
  extraction package, phenofit version 0.2.2, https://github.com/kongdd/phenofit


  \item Zhang, Q., Kong, D., Shi, P., Singh, V.P., Sun, P., 2018. Vegetation phenology 
  on the Qinghai-Tibetan Plateau and its response to climate change (1982–2013). 
  Agric. For. Meteorol. 248, 408–417. https://doi.org/10.1016/j.agrformet.2017.10.026
}
}
