% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAI_proc_spline.R
\name{LAI_proc_spline}
\alias{LAI_proc_spline}
\title{Interpolates LAI into daily values}
\usage{
LAI_proc_spline(Site)
}
\arguments{
\item{Site}{The Site ID}
}
\description{
This function generates a continuous series of LAI from downloaded
from the lpdaac AppEEARS download portal through spline interpolation. Processing
steps include:

\itemize{
  \item Savitsky-Golay filtering
  \item Fit a double logistic function to extract a continuous series of LAI
}
}
