% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_standardized.R
\name{light_standardized}
\alias{light_standardized}
\title{Generates a standardized set of light validation data for StreamPULSE sites}
\usage{
light_standardized(read_dir, save_dir, Site, Lat, Lon, site_crs)
}
\arguments{
\item{read_dir}{The read directory for downloaded files. For example, "C:/}

\item{save_dir}{The save directory for files to be placed in. For example, "C:/}

\item{Site}{The site name, for example "FL_ICHE2700"}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}

\item{site_crs}{The coordinate reference system of the points, preferably designated
as an EPSG code. For example, the most common geographic system is WGS84 and its EPSG 
code is 4326.}
}
\value{
Returns a standardized file of light data
}
\description{
This function takes observed light, in either Lux or PPFD, from
StreamPULSE sites and generates a standardized set of validation data for
comparing estimated light.
}
