% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_driver.R
\name{make_driver}
\alias{make_driver}
\title{Makes light model driver files}
\usage{
make_driver(
  site_locs,
  nldas_prepped,
  modis_prepped,
  write_output = FALSE,
  save_dir = NULL
)
}
\arguments{
\item{site_locs}{A table with Site_ID, Lat, and Lon, and the coordinate
reference system designated as an EPSG code. For example, the most common 
geographic system is WGS84 and its EPSG code is 4326}

\item{nldas_prepped}{Output from the nldas_prep function}

\item{modis_prepped}{Output from the appeears_prep function}

\item{write_output}{Binary indicating whether to write each individual driver
file to disk. Default value is FALSE.}

\item{save_dir}{Optional parameter when write_output = TRUE. The save directory 
for files to be placed in. For example, "C:/}
}
\value{
Returns a driver time series
}
\description{
This function compiles NLDAS light data and MODIS LAI
}
