% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLDAS_DL.R
\name{NLDAS_DL}
\alias{NLDAS_DL}
\title{Downloads NLDAS light data}
\usage{
NLDAS_DL(save_dir, Site_ID, Lat, Lon, startDate)
}
\arguments{
\item{save_dir}{The save directory for files to be placed in. For example, "C:/myfolder}

\item{Site_ID}{The site ID, for example "FL_ICHE2700"}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}

\item{startDate}{The starting date for the download (YYYY-MM-DD)}
}
\value{
Returns a time series of incoming shortwave solar radiation from the start
date to the most recent available data
}
\description{
This function downloads NLDAS incoming shortwave radiation data
(w m-2) for a given Latitude and Longitude.
}
