% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLDAS_DL_bulk.R
\name{NLDAS_DL_bulk}
\alias{NLDAS_DL_bulk}
\title{Downloads NLDAS light data}
\usage{
NLDAS_DL_bulk(save_dir, site_locs, startDate)
}
\arguments{
\item{save_dir}{The save directory for files to be placed in. For example, "C:/myfolder}

\item{site_locs}{A table with Site_ID, Lat, and Lon, and startDate}

\item{statDate}{An optional parameter. By default, if nothing is provided the function
assumes that site_locs has a column that contains startDate. Alternatively, a single
startDate can be provided as an argument for the download (YYYY-MM-DD).}
}
\value{
Returns a time series of incoming shortwave solar radiation from the start
date to the most recent available data
}
\description{
This function downloads NLDAS incoming shortwave radiation data
(w m-2). When using the NLDAS_DL function to download data from many sites, sometimes
sites will fail to download data (possibly due to volume of requests). This function
adds in an additional check to continually attempt to download data for any missing
locations until no more locations can be successfully downloaded. For example, sites
outside of the U.S. do not have NLDAS data and therefore will never download.
}
