% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldas_get.R
\name{nldas_get}
\alias{nldas_get}
\title{Downloads NLDAS light data}
\usage{
nldas_get(
  Site_ID,
  Lat,
  Lon,
  startDate,
  endDate = Sys.Date(),
  access_token,
  save_dir
)
}
\arguments{
\item{Site_ID}{The site ID, for example "NC_NHC"}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}

\item{startDate}{The starting date for the download (YYYY-MM-DD)}

\item{endDate}{The ending date for the download (YYYY-MM-DD). Default value is 
the current date via Sys.Date().}

\item{access_token}{An Earthdata access token. Requires an Earthdata profile
https://urs.earthdata.nasa.gov/. From your account page click generate token
and copy the token into a local file, which will need to be read in as your
access_token. e.g. access_token <- readLines("path/token.txt")}

\item{save_dir}{The save directory for files to be placed in. For example, "C:/myfolder}
}
\value{
Returns a time series of incoming shortwave solar radiation from the start
date to the end date.
}
\description{
This function downloads NLDAS incoming shortwave radiation data
(w m-2) for a given Latitude and Longitude.
}
