% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldas_get_bulk.R
\name{nldas_get_bulk}
\alias{nldas_get_bulk}
\title{Bulk downloads NLDAS light data for multiple sites}
\usage{
nldas_get_bulk(site_locs, startDate, endDate, access_token, save_dir)
}
\arguments{
\item{site_locs}{A table with Site_ID, Lat, and Lon. Optionally, also startDate and endDate.}

\item{startDate}{An optional parameter. By default, if nothing is provided the function
assumes that site_locs has a column that contains startDate. Alternatively, a single
startDate can be provided as an argument for the download (YYYY-MM-DD).}

\item{endDate}{An optional parameter. By default, if nothing is provided the function
assumes that site_locs has a column that contains endDate. Alternatively, a single
endDate can be provided as an argument for the download (YYYY-MM-DD).}

\item{access_token}{An Earthdata access token. Requires an Earthdata profile
https://urs.earthdata.nasa.gov/. From your account page click generate token
and copy the token into a local file, which will need to be read in as your
access_token. e.g. access_token <- readLines("path/token.txt")}

\item{save_dir}{The save directory for files to be placed in. For example, "C:/myfolder}
}
\value{
Returns a time series of incoming shortwave solar radiation from the start
date to the most recent available data
}
\description{
This function downloads NLDAS incoming shortwave radiation data
(w m-2). When using the nldas_get function to download data from many sites, sometimes
sites may fail to download data (possibly due to volume of requests). This function
adds in an additional check to continually attempt to download data for any missing
locations until no more locations can be successfully downloaded. For example, sites
outside of the U.S. do not have NLDAS data and therefore will never download.

The startDate and endDate arguments may either both be specified in the function,
or provided as columns named "startDate" and "endDate" in the site_locs table.
}
