% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldas_prep.R
\name{nldas_prep}
\alias{nldas_prep}
\title{Compiles downloaded NLDAS data in preparation for making driver files}
\usage{
nldas_prep(read_dir, write_output = FALSE, save_dir = NULL)
}
\arguments{
\item{read_dir}{The read directory for downloaded files. For example, "C:/myfolder}

\item{write_output}{Logical value indicating whether to write each individual driver
file to disk. Default value is FALSE.}

\item{save_dir}{Optional parameter when write_output = TRUE. The save directory 
for files to be placed in. For example, "C:/}
}
\value{
Returns a list of downloaded NLDAS data, named by Site_ID
}
\description{
This function compiles all of the downloaded NLDAS data into
a list, with each element named with the Site_ID. This step is in preparation
for making driver files using the make_driver function.
}
