% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_ts.R
\name{pad_ts}
\alias{pad_ts}
\title{Pad a timeseries with a year at the start and the end}
\usage{
pad_ts(ts, t_res)
}
\arguments{
\item{ts}{timeseries}

\item{t_res}{The temporal resolution of the LAI data}
}
\value{
Returns a timeseries with an additional year added to the start and
end of the timeseries
}
\description{
This function pads a timeseries with a year of data at the start
and the end of the timeseries. One year of LAI data is added to the start and end
of the timeseries based on median daily LAI for the timeserires. This is done
so that a smoothed seasonal cycle can be retrieved from the timeseries using
the phenofit R package for the entire timeseries. .
}
