% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_weights.R
\name{qc_weights}
\alias{qc_weights}
\title{Assigns weights to LAI data based on the 5-level QC score}
\usage{
qc_weights(SCF_QC, wmin = 0.2, wmid = 0.5, wmax = 1)
}
\arguments{
\item{SCF_QC}{The 5-level QC score (FparLai_QC_SCF_QC)}

\item{wmin}{The minimum weight for a data point}

\item{wmid}{The middle weight for a data point}

\item{wmax}{The maximum weight for a data point (best quality)}
}
\value{
Returns weights based on the 5-level QC score
}
\description{
This function unpacks downloaded MODIS LAI data from the lpdaac
AppEEARS download portal into a list with each item in the list representing
a single site.
}
