% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simard2011.R
\docType{data}
\name{simard2011}
\alias{simard2011}
\title{LiDAR derived estimates of tree heights from Simard et al. (2011)}
\format{GeoTIFF}
\usage{
data(simard2011)
}
\description{
Contains 1km resolution estimates of tree height produced by
Simard et al. (2011) Mapping forest canopy height globally with spaceborne
lidar. The data can be accessed from (https://webmap.ornl.gov/ogc/dataset.jsp?ds_id=10023)
}
\keyword{datasets}
