% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_world.R
\docType{data}
\name{tz_world}
\alias{tz_world}
\title{A shapefile of world timezones}
\format{Shapefile}
\usage{
data(tz_world)
}
\description{
A shapefile containing the boundaries of timezones across the world downloaded
from (http://efele.net/maps/tz/world/) that was created by Eric Muller.
The specific file used is tz_world.zip.
}
\keyword{datasets}
