% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppEEARS_proc.R
\name{AppEEARS_proc}
\alias{AppEEARS_proc}
\title{Filters and gap-fills downloaded MODIS AppEEARS data}
\usage{
AppEEARS_proc(Site, proc_type)
}
\arguments{
\item{Site}{The site name}

\item{proc_type}{either "spline" (for a spline) or "dbl_log" (for a double logistic function)}
}
\description{
This function filters and gap-fills downloaded MODIS LAI data from
the lpdaac AppEEARS download portal. Processing steps include:

\itemize{
  \item Savitsky-Golay filtering
  \item Interpolation using either a spline or double logistic function
}
}
