% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_height.R
\name{extract_height}
\alias{extract_height}
\title{Get LiDAR derived tree heights based on Latitude and Longitude}
\usage{
extract_height(Site_ID, Lat, Lon)
}
\arguments{
\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}

\item{Site}{The site ID}
}
\value{
Site ID, Latitude, Longitude, and tree height
}
\description{
Extract canopy height from the LiDAR derived map of
tree heights produced by Simard et al. 2011 (asdf)
}
