% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_driver.R
\name{make_driver}
\alias{make_driver}
\title{Makes light model driver files}
\usage{
make_driver(site_locs, NLDAS_processed, MOD_processed, save_dir)
}
\arguments{
\item{site_locs}{A table with Site_ID, Lat, and Lon}

\item{NLDAS_processed}{Output from the NLDAS_proc function}

\item{MOD_processed}{Output from the}

\item{save_dir}{The save directory for files to be placed in. For example, "C:/}
}
\value{
Returns a driver time series
}
\description{
This function compiles NLDAS light data and MODIS LAI
}
